Story/requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
dj-rest-auth==6.0.0
Django==5.1
django-allauth==64.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
gunicorn==23.0.0
idna==3.8
inflection==0.5.1
packaging==24.1
PyJWT==2.9.0
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
setuptools==73.0.1
sqlparse==0.5.1
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
